Move homepage content, image macro, svg sprite
The homepage content was under the templates for some reason. Now it is with the rest of the content.

Introduces a macro for images. Images have been moved to Cloudinary from Google Photos, so this should enable more covenient references and responsive features (srcset, sizes) in the future.

In order to support macros in non-template contexts, the templating plugin (metalsmith-templates) was upgraded. It had been deprecated in favor of metalsmith-layouts, which matches the existing use case, and metalsmith-in-place, which uses JSTransformer to enable templating of arbitrary files by naming convention. The in-place templating now handles cases where we want content to use macros.

There was also an issue with SVG spriting where things were left in a broken condition after attempting to automatically discover SVG icon uses within the site content. While it seems like that would be useful, there are other things I'd like to get to first. I added an icons.yaml file that acts as a manifest of SVG icons to extract from IcoMoon. This is probably a satisfactory solution anyway, since those things don't change often.