Fix footer commit
beautifulhugo [moved away from using the environment variable] and to the built-in `GitInfo` template variable.

Enabling `GitInfo` also had a couple of side-effects, like showing non-content modifications (i.e. frontmatter-only) to posts as "updates" when presenting the post's date. We change the `lastmod` behavior to exclude Git as a source of this information in order to avoid this unexpectedness.

[moved away from using the environment variable]: https://github.com/halogenica/beautifulhugo/pull/235